hw2/requirements.txt
pyspark==3.5.1
pywikibot==9.5.0
sseclient==0.0.27
kafka_python==2.0.2